Acomodando y modularizando estilos
