+++ src/Pages/Cart/Cart.tsx
@@ -3,7 +3,7 @@ import { useCart } from '@/Contexts/CartContext';
 import { useNavigate } from 'react-router-dom';
 import { handleRemove, handleClearCart } from '@/Utils/utils';
 
-import '@/Styles/AppBar.css';
+import './Cart.module.css';
 
 const Cart: React.FC = () => {
   const { state, dispatch } = useCart();
@@ -17,27 +17,14 @@ const Cart: React.FC = () => {
     <div style={{ textAlign: 'center', padding: '20px' }}>
       <h2 className="carritotitle">Tu carrito</h2>
       {state.items.length === 0 ? (
-        <div
-          style={{
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            justifyContent: 'center',
-            height: '60vh',
-          }}
-        >
-          <h2 style={{ fontSize: '2rem', marginBottom: '20px', color: '#555' }}>
+        <div className='CartContent'>
+          <h2>
             Tu carrito está vacío
           </h2>
           <img
             src="/assets/gatitriste.png" // URL relativa para imágenes en "public"
             alt="Carrito vacío"
-            style={{
-              width: '200px',
-              height: '200px',
-              objectFit: 'contain',
-              marginBottom: '20px',
-            }}
+            className='CartEmpty'
           />
         </div>
       ) : (
