+++ src/Pages/ProductList/ProductList.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import { fetchProducts } from '@/Services/productoService';
-import { Product, ProductListProps } from '../Types/types';
+import { Product, ProductListProps } from '../../Types/types';
 
 const ProductList: React.FC<ProductListProps> = ({ searchTerm, selectedCategory }) => {
   const [products, setProducts] = useState<Product[]>([]);
