--- src/Styles/App.css
@@ -1,110 +1,0 @@
-HEAD
-.app-bar {
-  background-color: #0052cc;
-  padding: 1rem;
-  display: flex;
-  justify-content: center; /* Centra el contenido horizontalmente */
-  align-items: center;
-  width: 100%;
-}
-
-.app-bar-content {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  max-width: 1280px;
-}
-
-.logo-section {
-  display: flex;
-  align-items: center;
-}
-
-.title-image {
-  width: 200px;
-  height: auto;
-  margin-right: 100px;
- margin-left: 200px;
-}
-
-.navbar-search {
-  flex-grow: 1;
-  text-align: center;
-}
-
-.navbar-search input {
-  padding: 0.5rem;
-  border-radius: 5px;
-  border: 1px solid #ccc;
-  width: 300px;
-}
-
-.navbar-icons {
-  display: flex;
-  gap: 1rem;
-}
-
-.icon {
-  height: 24px;
-  cursor: pointer;
-}
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem;
-  background-color: #0052cc; /* Cambia esto según el color de fondo que desees */
-  color: white;
-}
-
-.navbar-logo {
-  display: flex;
-  align-items: center;
-}
-
-.logo {
-  height: 50px;
-  margin-right: 10px;
-}
-
-.navbar-search {
-  flex-grow: 1;
-  text-align: center;
-}
-
-.navbar-search input {
-  padding: 0.5rem;
-  border-radius: 4px;
-  background-color: white;
-  border: 1px solid #000;
-  width: 300px; /* Ajusta el ancho según lo que necesites */
-}
-
-.navbar-icons {
-  list-style: none;
-  display: flex;
-  gap: 1.5rem;
-  margin: 0;
-}
-
-.icon {
-  height: 24px;
-  cursor: pointer;
-}
-
-.navbar-icons a {
-  color: white;
-  text-decoration: none;
-
-}
-form {
-  margin-top: 160px;
-}
-
-
-form {
-  display: block;
-  margin-top: 0em;
-  unicode-bidi: isolate;
-}
